requirements/base.txt
# SHA1:35e0f85a0fc1d6da12e1fdbbd74e09e7fb4a784c
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
grpcio==1.36.1
    # via -r requirements/base.in
protobuf==3.6.1
    # via -r requirements/base.in
six==1.15.0
    # via
    #   grpcio
    #   protobuf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
